+++ fractal_tasks_core/cellpose_segmentation.py
@@ -627,7 +627,7 @@ if __name__ == "__main__":
         level: int
         relabeling: bool = True
         anisotropy: Optional[float] = None
-        diameter_level0: float
+        diameter_level0: Optional[float]
         cellprob_threshold: Optional[float]
         flow_threshold: Optional[float]
         ROI_table_name: Optional[str]