+++ fractal_tasks_core/copy_ome_zarr.py
@@ -188,9 +188,10 @@ def copy_ome_zarr(
 
 if __name__ == "__main__":
     from pydantic import BaseModel
+    from pydantic import Extra
     from fractal_tasks_core._utils import run_fractal_task
 
-    class TaskArguments(BaseModel):
+    class TaskArguments(BaseModel, extra=Extra.forbid):
         input_paths: Sequence[str]
         output_path: str
         metadata: Dict[str, Any]