+++ fractal_tasks_core/maximum_intensity_projection.py
@@ -132,9 +132,10 @@ def maximum_intensity_projection(
 
 if __name__ == "__main__":
     from pydantic import BaseModel
+    from pydantic import Extra
     from fractal_tasks_core._utils import run_fractal_task
 
-    class TaskArguments(BaseModel):
+    class TaskArguments(BaseModel, extra=Extra.forbid):
         input_paths: Sequence[str]
         output_path: str
         metadata: Dict[str, Any]