+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# 0.8.0
+# Unreleased
 
 * **(major)** Update task interface: remove filename extension from `input_paths` and `output_path` for all tasks, and add new arguments `(image_extension,image_glob_pattern)` to `create_ome_zarr` task (\#323).
+* Fix minor bugs in task arguments (\#329).
 
 # 0.7.5